+++ library/src/lapack/roclapack_syevd_heevd.hpp
@@ -1,0 +1,178 @@
+/************************************************************************
+ * Derived from the BSD3-licensed
+ * LAPACK routine (version 3.7.0) --
+ *     Univ. of Tennessee, Univ. of California Berkeley,
+ *     Univ. of Colorado Denver and NAG Ltd..
+ *     December 2016
+ * Copyright (c) 2021 Advanced Micro Devices, Inc.
+ * ***********************************************************************/
+
+#pragma once
+
+#include "../auxiliary/rocauxiliary_ormtr_unmtr.hpp"
+#include "../auxiliary/rocauxiliary_stedc.hpp"
+#include "roclapack_syev_heev.hpp"
+
+/** Helper to calculate workspace sizes **/
+template <bool BATCHED, typename T, typename S>
+void rocsolver_syevd_heevd_getMemorySize(const rocblas_evect evect,
+                                         const rocblas_fill uplo,
+                                         const rocblas_int n,
+                                         const rocblas_int batch_count,
+                                         size_t* size_scalars,
+                                         size_t* size_work1,
+                                         size_t* size_work2,
+                                         size_t* size_work3,
+                                         size_t* size_tmptau_W,
+                                         size_t* size_tau,
+                                         size_t* size_workArr)
+{
+    // if quick return, set workspace to zero
+    if(n <= 1 || batch_count == 0)
+    {
+        *size_scalars = 0;
+        *size_work1 = 0;
+        *size_work2 = 0;
+        *size_work3 = 0;
+        *size_tmptau_W = 0;
+        *size_tau = 0;
+        *size_workArr = 0;
+        return;
+    }
+
+    size_t unused;
+    size_t w11 = 0, w12 = 0, w13 = 0;
+    size_t w21 = 0, w22 = 0, w23 = 0;
+    size_t w31 = 0, w32 = 0;
+    size_t t1 = 0, t2 = 0;
+
+    // requirements for tridiagonalization (sytrd/hetrd)
+    rocsolver_sytrd_hetrd_getMemorySize<T, BATCHED>(n, batch_count, size_scalars, &w11, &w21, &t1,
+                                                    &unused);
+
+    if(evect == rocblas_evect_original)
+    {
+        // extra requirements for computing eigenvalues and vectors (stedc)
+        rocsolver_stedc_getMemorySize<BATCHED, T, S>(evect, n, batch_count, &w12, &w22, &w31,
+                                                     &unused);
+
+        // extra requirements for ormtr/unmtr
+        rocsolver_ormtr_unmtr_getMemorySize<T, BATCHED>(rocblas_side_left, uplo, n, n, batch_count,
+                                                        &unused, &w13, &w23, &w32, &unused);
+
+        *size_work3 = std::max(w31, w32);
+    }
+    else
+    {
+        // extra requirements for computing only the eigenvalues (sterf)
+        rocsolver_sterf_getMemorySize<T>(n, batch_count, &w12);
+
+        *size_work3 = 0;
+    }
+
+    // size of array for temporary matrix products
+    t2 = sizeof(T) * n * n * batch_count;
+
+    // get max values
+    *size_work1 = std::max({w11, w12, w13});
+    *size_work2 = std::max({w21, w22, w23});
+    *size_tmptau_W = std::max(t1, t2);
+
+    // size of array for temporary householder scalars
+    *size_tau = sizeof(T) * n * batch_count;
+
+    // size of array of pointers to workspace
+    if(BATCHED)
+        *size_workArr = 2 * sizeof(T*) * batch_count;
+    else
+        *size_workArr = 0;
+}
+
+template <bool BATCHED, bool STRIDED, typename T, typename S, typename W>
+rocblas_status rocsolver_syevd_heevd_template(rocblas_handle handle,
+                                              const rocblas_evect evect,
+                                              const rocblas_fill uplo,
+                                              const rocblas_int n,
+                                              W A,
+                                              const rocblas_int shiftA,
+                                              const rocblas_int lda,
+                                              const rocblas_stride strideA,
+                                              S* D,
+                                              const rocblas_stride strideD,
+                                              S* E,
+                                              const rocblas_stride strideE,
+                                              rocblas_int* info,
+                                              const rocblas_int batch_count,
+                                              T* scalars,
+                                              void* work1,
+                                              void* work2,
+                                              void* work3,
+                                              T* tmptau_W,
+                                              T* tau,
+                                              T** workArr)
+{
+    ROCSOLVER_ENTER("syevd_heevd", "evect:", evect, "uplo:", uplo, "n:", n, "shiftA:", shiftA,
+                    "lda:", lda, "bc:", batch_count);
+
+    // quick return
+    if(batch_count == 0)
+        return rocblas_status_success;
+
+    hipStream_t stream;
+    rocblas_get_stream(handle, &stream);
+
+    rocblas_int blocksReset = (batch_count - 1) / BLOCKSIZE + 1;
+    dim3 gridReset(blocksReset, 1, 1);
+    dim3 threads(BLOCKSIZE, 1, 1);
+
+    // info = 0
+    hipLaunchKernelGGL(reset_info, gridReset, threads, 0, stream, info, batch_count, 0);
+
+    // quick return
+    if(n == 0)
+        return rocblas_status_success;
+
+    // quick return for n = 1 (scalar case)
+    if(n == 1)
+    {
+        hipLaunchKernelGGL(scalar_case<T>, gridReset, threads, 0, stream, evect, A, strideA, D,
+                           strideD, batch_count);
+        return rocblas_status_success;
+    }
+
+    // TODO: Scale the matrix
+
+    // reduce A to tridiagonal form
+    rocsolver_sytrd_hetrd_template(handle, uplo, n, A, shiftA, lda, strideA, D, strideD, E, strideE,
+                                   tau, n, batch_count, scalars, (T*)work1, (T*)work2, tmptau_W,
+                                   workArr);
+
+    if(evect != rocblas_evect_original)
+    {
+        // only compute eigenvalues
+        rocsolver_sterf_template<S>(handle, n, D, 0, strideD, E, 0, strideE, info, batch_count,
+                                    (rocblas_int*)work1);
+    }
+    else
+    {
+        constexpr bool ISBATCHED = BATCHED || STRIDED;
+        const rocblas_int ldw = n;
+        const rocblas_stride strideW = n * n;
+
+        rocsolver_stedc_template<false, ISBATCHED, T>(
+            handle, rocblas_evect_tridiagonal, n, D, 0, strideD, E, 0, strideE, tmptau_W, 0, ldw,
+            strideW, info, batch_count, work1, (S*)work2, (S*)work3, (S**)workArr);
+
+        rocsolver_ormtr_unmtr_template<BATCHED, STRIDED>(
+            handle, rocblas_side_left, uplo, rocblas_operation_none, n, n, A, shiftA, lda, strideA,
+            tau, n, tmptau_W, 0, ldw, strideW, batch_count, scalars, (T*)work1, (T*)work2,
+            (T*)work3, workArr);
+
+        // copy matrix product into A
+        const rocblas_int copyblocks = (n - 1) / 32 + 1;
+        hipLaunchKernelGGL(copy_mat<T>, dim3(copyblocks, copyblocks, batch_count), dim3(32, 32), 0,
+                           stream, n, n, tmptau_W, 0, ldw, strideW, A, shiftA, lda, strideA);
+    }
+
+    return rocblas_status_success;
+}